+++ lib/src/issel_info_field.dart
@@ -49,7 +49,7 @@ class IsselInfoField extends StatelessWidget {
           Expanded(
               child: Container(
             decoration: BoxDecoration(
-                color: valueBackColor ?? theme.scaffoldBackgroundColor,
+                color: valueBackColor ?? colorScheme.surfaceContainer,
                 borderRadius: BorderRadius.circular(10)),
             child: Center(
               child: Text(